src/lambda/chat_authorizer/requirements.txt
python-jose[cryptography]==3.3.0
boto3>=1.34.0
